# packages/contracts/openapi.yaml
openapi: 3.1.0
info:
  title: CCAT Practice App — Gateway API
  version: "1.0.0-phase0"
  description: >
    Normative /v1 API for the Concept Mastery CCAT Practice App. The Gateway is the ONLY
    application-data boundary exposed to the student app (Blueprint v10.0 §1.1, §33).
    Direct student Supabase/PostgREST access is prohibited (§41). All sensitive mutations
    use idempotency keys; all mutable Admin resources use expected-version/ETag concurrency
    (§22.4, §32.1). This document is Phase 0 and covers the endpoint surface of §32; request
    bodies for Admin write operations are represented by shared envelopes and will be
    expanded per-domain in Phase 1.
  contact:
    name: Concept Mastery
    email: admin@conceptmastery.ca
servers:
  - url: https://api.conceptmastery.ca/v1
    description: Production
  - url: https://staging-api.conceptmastery.ca/v1
    description: Staging (no real child PII, §38.1)
...
security:
  - studentSession: []
paths:
...
  /admin/rewards/adjust:
    post:
      tags: [admin-rewards]
      operationId: adminRewardAdjust
      security: [ { adminSession: [] } ]
      summary: Compensating XP/coin ledger adjustment (§19.3). Never overwrites history.
      parameters: [ { $ref: "#/components/parameters/IdempotencyKey" } ]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/RewardAdjustment" }
      responses:
        "201": { description: Adjustment ledgered }
        "403": { $ref: "#/components/responses/Error" }
...
components:
  securitySchemes:
    studentSession:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Student application session token, bound to the sole enrolled device (§5.4).
    adminSession:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Admin identity token (email/password + mandatory MFA, §22.1). Permissions loaded server-side.
  parameters:
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      schema: { type: string, maxLength: 200 }
      description: Required on sensitive mutations (§13, §32.1). Retries with the same key return the original result.
...
  responses:
    Error:
      description: Structured error envelope (§32.1)
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
...
  schemas:
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code: { type: string, description: Machine code, e.g. ACTIVE_SESSION_EXISTS }
            message: { type: string }
            request_id: { type: string }
            details: { type: object, additionalProperties: true }
...
    RewardAdjustment:
      type: object
      required: [student_id, kind, delta, reason, reference]
      properties:
        student_id: { type: string, format: uuid }
        kind: { type: string, enum: [xp, coins] }
        delta: { type: integer, description: May be negative; compensating entry (§19.3) }
        reason: { type: string }
        reference: { type: string }
...